requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
webdriver-manager
Appium-Python-Client